// frontend/src/types/types.ts
export interface ProductResponse {
    productId: number;
    productName: string;
    productCategory: string;
    price: number;
    review: number;
    imagePath: string;
    reviewsCount: number;
    
}

export interface FullPerfumeResponse extends ProductResponse {
    year: number;
    seller: string;
    
    
    description: string;
    type: string;
    file: any;
}

export interface HeaderResponse<T> {
    items: Array<T>;
    pagesCount: number;
    totalElements: number;
}

export interface UserOrdersRequest {
    email: string;
    page: number;
}

export interface PerfumesSearchRequest {
   
    searchtext: string;
    
}